Added Loggable mixin (#1)
* Added Loggable mixin

The approach is copied over from ecraft-web_api. The plan is to let that gem use this gem eventually, so we make the use case for "logging in a non-web app" more obvious. (Why would you want to depend on ecraft-web_api just to get logging? It doesn't really make sense.)

* Peer suggestions: Added README.md to describe usage

Also added more specs.