Release 1.0.0
This feels like a big jump, but: I don't like the idea of letting uxFactory Server 8.x and its ecosystem depend on a bunch of prerelease gems. Hence, jumping to 1.0 _very_ quickly in this case. It simply means: we must keep the public API now with diligence, to avoid breaking the SemVer contract with our users.