Bug fix: CustomFormatter#formatted_message was broken under certain circumstances
When `message` was neither a `String` nor an `Exception`, the code was broken. It could never work in that case.

Sadly, the lack of _specs_ is a bad fit for a dynamic language. This is actually an example of the the small but subtle bugs that static analysis done by e.g. a _compiler_ would have spotted long before I made a release of the gem including the bug... No, I'm not bitter, not at all. ;)